requirements.txt
arxiv
openai
pydantic
python-dotenv
requests
scrapy
tqdm
